+++ contracts/protocol/configuration/StrategyManager.sol
@@ -207,7 +207,7 @@ contract StrategyManager is IStrategyManager, Modifiers {
     ) internal view returns (bytes[] memory _codes) {
         DataTypes.StrategyStep[] memory _strategySteps = _getStrategySteps(_investStrategyhash);
         uint8 _subStepCounter = 0;
-        for (uint256 _i = 0; _i < uint256(_strategySteps.length); _i++) {
+        for (uint256 _i = 0; _i < _strategySteps.length; _i++) {
             if (_strategySteps[_i].isBorrow) {
                 if (_stepIndex == _subStepCounter) {
                     address _liquidityPool = _strategySteps[_i].pool;
@@ -250,7 +250,7 @@ contract StrategyManager is IStrategyManager, Modifiers {
                     _codes = IAdapter(_adapter).getDepositAllCodes(_vault, _underlyingTokens, _liquidityPool);
                     break;
                 } // deposit at ith step
-                if (_stepIndex == (_subStepCounter + 1) && _i == uint256(_strategySteps.length - 1)) {
+                if (_stepIndex == (_subStepCounter + 1) && _i == _strategySteps.length - 1) {
                     address _liquidityPool = _strategySteps[_i].pool;
                     address _adapter = registryContract.getLiquidityPoolToAdapter(_liquidityPool);
                     address[] memory _underlyingTokens = new address[](1);
@@ -274,8 +274,8 @@ contract StrategyManager is IStrategyManager, Modifiers {
     ) internal view returns (bytes[] memory _codes) {
         DataTypes.StrategyStep[] memory _strategySteps = _getStrategySteps(_investStrategyhash);
         uint256 _subStepCounter = _stepCount - 1;
-        for (uint256 _i = 0; _i < uint256(_strategySteps.length); _i++) {
-            uint256 _iterator = uint256(_strategySteps.length) - 1 - _i;
+        for (uint256 _i = 0; _i < _strategySteps.length; _i++) {
+            uint256 _iterator = _strategySteps.length - 1 - _i;
             if (_strategySteps[_iterator].isBorrow) {
                 address _outputToken = _strategySteps[_iterator].outputToken;
                 if (_stepIndex == _subStepCounter) {
@@ -311,7 +311,7 @@ contract StrategyManager is IStrategyManager, Modifiers {
                     address _adapter = registryContract.getLiquidityPoolToAdapter(_strategySteps[_iterator].pool);
                     address[] memory _underlyingTokens = new address[](1);
                     _underlyingTokens[0] = _underlyingToken;
-                    _codes = (_iterator == uint256(_strategySteps.length) - 1 &&
+                    _codes = (_iterator == _strategySteps.length - 1 &&
                         IAdapter(_adapter).canStake(_strategySteps[_iterator].pool))
                         ? IAdapter(_adapter).getUnstakeAndWithdrawAllCodes(
                             _vault,
@@ -376,11 +376,11 @@ contract StrategyManager is IStrategyManager, Modifiers {
         address _underlyingToken,
         bytes32 _investStrategyhash
     ) internal view returns (uint256 _balance) {
-        uint256 _steps = uint256(_getStrategySteps(_investStrategyhash).length);
+        uint256 _steps = _getStrategySteps(_investStrategyhash).length;
         DataTypes.StrategyStep[] memory _strategySteps = _getStrategySteps(_investStrategyhash);
         _balance = 0;
         uint256 _outputTokenAmount = _balance;
-        for (uint256 _i = 0; _i < uint256(_steps); _i++) {
+        for (uint256 _i = 0; _i < _steps; _i++) {
             uint256 _iterator = _steps - 1 - _i;
             address _liquidityPool = _strategySteps[_iterator].pool;
             address _adapter = registryContract.getLiquidityPoolToAdapter(_liquidityPool);
@@ -416,7 +416,7 @@ contract StrategyManager is IStrategyManager, Modifiers {
 
     function _getHarvestRewardStepsCount(bytes32 _investStrategyhash) internal view returns (uint8) {
         DataTypes.StrategyStep[] memory _strategySteps = _getStrategySteps(_investStrategyhash);
-        uint256 _lastStepIndex = uint256(_strategySteps.length) - 1;
+        uint256 _lastStepIndex = _strategySteps.length - 1;
         address _lastStepLiquidityPool = _strategySteps[_lastStepIndex].pool;
         address _lastStepOptyAdapter = registryContract.getLiquidityPoolToAdapter(_lastStepLiquidityPool);
         if (IAdapter(_lastStepOptyAdapter).getRewardToken(_lastStepLiquidityPool) != address(0)) {
@@ -427,7 +427,7 @@ contract StrategyManager is IStrategyManager, Modifiers {
 
     function _getClaimRewardStepsCount(bytes32 _investStrategyhash) internal view returns (uint8) {
         DataTypes.StrategyStep[] memory _strategySteps = _getStrategySteps(_investStrategyhash);
-        uint256 _lastStepIndex = uint256(_strategySteps.length) - 1;
+        uint256 _lastStepIndex = _strategySteps.length - 1;
         address _lastStepLiquidityPool = _strategySteps[_lastStepIndex].pool;
         address _lastStepOptyAdapter = registryContract.getLiquidityPoolToAdapter(_lastStepLiquidityPool);
         if (IAdapter(_lastStepOptyAdapter).getRewardToken(_lastStepLiquidityPool) != address(0)) {
@@ -441,14 +441,14 @@ contract StrategyManager is IStrategyManager, Modifiers {
             return uint8(0);
         }
         DataTypes.StrategyStep[] memory _strategySteps = _getStrategySteps(_investStrategyhash);
-        uint256 _strategyStepCount = uint256(_strategySteps.length);
+        uint256 _strategyStepCount = _strategySteps.length;
         uint256 _lastStepIndex = _strategyStepCount - 1;
         address _lastStepLiquidityPool = _strategySteps[_lastStepIndex].pool;
         address _lastStepOptyAdapter = registryContract.getLiquidityPoolToAdapter(_lastStepLiquidityPool);
         if (IAdapter(_lastStepOptyAdapter).canStake(_lastStepLiquidityPool)) {
             return (_strategyStepCount + 1);
         }
-        for (uint256 i = 0; i < uint256(_strategySteps.length); i++) {
+        for (uint256 i = 0; i < _strategySteps.length; i++) {
             if (_strategySteps[i].isBorrow) {
                 _strategyStepCount++;
             }
@@ -461,8 +461,8 @@ contract StrategyManager is IStrategyManager, Modifiers {
             return uint256(0);
         }
         DataTypes.StrategyStep[] memory _strategySteps = _getStrategySteps(_investStrategyhash);
-        uint256 _steps = uint256(_strategySteps.length);
-        for (uint256 _i = 0; _i < uint256(_strategySteps.length); _i++) {
+        uint256 _steps = _strategySteps.length;
+        for (uint256 _i = 0; _i < _strategySteps.length; _i++) {
             if (_strategySteps[_i].isBorrow) {
                 _steps++;
             }
@@ -482,7 +482,7 @@ contract StrategyManager is IStrategyManager, Modifiers {
             uint256 _treasurySharesLength = _treasuryShares.length;
             _treasuryCodes = new bytes[](_treasurySharesLength.add(1));
             if (_treasurySharesLength > 0) {
-                for (_i = 0; _i < uint256(_treasurySharesLength); _i++) {
+                for (_i = 0; _i < _treasurySharesLength; _i++) {
                     if (_treasuryShares[_i].treasury != address(0)) {
                         uint256 _share = _treasuryShares[_i].share;
                         uint256 _treasuryAccountFee = ((_redeemAmountInToken).mul(_share)).div(10000);
@@ -491,7 +491,7 @@ contract StrategyManager is IStrategyManager, Modifiers {
                             abi.encodeWithSignature(
                                 "transfer(address,uint256)",
                                 _treasuryShares[_i].treasury,
-                                uint256(_treasuryAccountFee)
+                                _treasuryAccountFee
                             )
                         );
                         _fee = _fee.add(_treasuryAccountFee);
